requirements.txt
# Macro Event Impact Prediction System
# Core dependencies

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# API clients
requests>=2.31.0
fredapi>=0.5.1
yfinance>=0.2.30
alpha-vantage>=2.3.1

# Web scraping for economic calendar
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Machine learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Database
sqlalchemy>=2.0.0
sqlite3-api>=2.0.0

# Async and scheduling
aiohttp>=3.9.0
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
